requirements.txt
aiogram
aiomysql
python-dotenv
apscheduler
cryptography
aiohttp